page

+++ src/ddd/dto.rs
@@ -9,7 +9,7 @@ pub struct MultipartFile {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct PageReq<T: Clone> {
-    pub page_no: Option<u32>,
-    pub page_size: Option<u16>,
+    pub page_no: Option<i64>,
+    pub page_size: Option<i32>,
     pub req: Option<T>,
 }
